Fix for transferlogic
First the problem was that if the object was transfered to a new chunk that wasnt updated yet that it got updated again and if the move was far enough that it even could happen that the object updated infinite times.

To fix this I implemented a Queue that transferes the Objects after each updatecycle. Then a new problem arose where at every second chunk transfer the game crashed. The problem was that before the Queue I could get the old and new chunk of the gameobject simply by looking where it was before and where it was now and using the methods in FinderHash but now it used the getChunkByObject Method that  returned which chunk the GameObject was registered in. For some reason all the time I got a nullptr back but after searching a long time I saw that when creating Chunks with getChunksInReach in the FinderHash they only got added to the FinderHash List but not to the chunks HashSet so the problem wasnt that the gameobjects didnt get assigned to the right chunk but that the chunk in which the gameobject now was, was not in the list I was searching in.

Yes too long sorry...