Layer List added
Added a List that keeps the elements in order. In the future it could be implemented that the list gets sorted by the layers so that lower layers get painted first but the main problem is that each chunk paints all of its objects, than the next and so forth, but if an object with a higher layer is painted last in a chunk and overlaps the chunks edge, every element of the other chunk will be painted on top of it regardless of the layer...

Flappy Bird can be used again like normal.